pyproject.toml
[project]
name = "audio-video-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "torch==2.7.1",
  "torchvision==0.22.1",
  "torchaudio==2.7.1",
  "funasr==1.2.7",
  "langchain-core==1.0.2",
  "python-dotenv==1.2.1",
  "pydub==0.25.1",
  "pydantic==2.12.3",
  "pydantic-core==2.41.4",
  "nvitop==1.5.3",
  "openai==2.6.1",
  "loguru==0.7.3",
  "faster-whisper>=1.2.1",
  "dashscope>=1.25.3",
  "langchain-openai>=1.0.3",
  "moviepy==1.0.3",
  "langchain>=1.0.4",
  "flask==3.0.3",
  "peewee==3.17.8",
  "redis==5.2.1",
  "flask-socketio==5.5.0",
  "celery==5.4.0",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu118"},
]
torchvision = [
  { index = "pytorch-cu118"},
]
torchaudio = [
  { index = "pytorch-cu118"},
]

